Add files via upload
I did not fully understand the following section but I hope everything else was according to the exercise :)! 

'Change your external link to an internal link by changing the href attribute to "#footer" and the text from Github to Jump to Bottom. Then add an id attribute with a value of footer to the <footer> element at the bottom of the page.'